StarsFilter added

+++ nodes/mappers/stars/rabbitmq/records.go
@@ -4,7 +4,7 @@ type FullReview struct {
 	ReviewId 		string 						`json:"review_id",omitempty`
 	UserId 			string 						`json:"user_id",omitempty`
 	BusinessId 		string 						`json:"business_id",omitempty`
-	Stars 			int 						`json:"stars",omitempty`
+	Stars 			float32 					`json:"stars",omitempty`
 	Useful			int 						`json:"useful",omitempty`
 	Funny 			int 						`json:"funny",omitempty`
 	Cool			int 						`json:"cool",omitempty`
@@ -14,5 +14,5 @@ type FullReview struct {
 
 type StarsData struct {
 	UserId 			string 						`json:"user_id",omitempty`
-	Stars 			int 						`json:"stars",omitempty`
+	Stars 			float32 					`json:"stars",omitempty`
 }
